Saved fights now maintain initiative order.  Passes tests.
The initiative order of two characters with same basic move is random and USED
to be re-rolled each time the fight was entered (so, if the fight was saved
and re-run, the initiative was rerolled) thereby potentially changing the
initiative order of the fight.  That's now fixed.

Playback seems to work for 'do_nothing' rounds.