Added code (commented out) to remove JSON items not in GCS.  Passes tests.
When updating a creature, some items in the JSON may not be in the GCS.
In some cases, it's good to remove those items but, in other cases,
those items should not be removed.  Code was added to remove them but
its been commented out.

Update character asks to remove items not in GCS.