Puts an item's capabilities with its type in the JSON.  Passes tests.
Weapons and armor, in the JSON, each have capabilities that used to be
included in the item's areas.  Now, an armor's dr is stored with it's
type (i.e., 'armor') and a weapon's capabilites are stored with the
weapon.  That allows a knife to be either thrown _or_ stabbed and the
skills and damage are stored with the 'thrown weapon' or 'thrust weapon'
type (as appropriate).

|type| is now a dict.  The type of an item, say 'armor', is the key and
the characteristics, say 'dr', is the value (which is also a dict).
Ultimately, this should allow a knife to be either thrown or stabbed or
slashed with all their information organized by the type of weapon.

Handles stuff like this (instead of 'melee weapon'):

    "swung weapon": {"damage": {"st": "sw", "plus": -2, "type": "cut"},
                     "skill": {"DR": -4, "Knife": 0},
    "thrust weapon": {"damage": {"st": "thr", "plus": 0, "type": "imp"},
                      "skill": {"DR": -4, "Knife": 0}

Modifies code to, among other things:
        * MainWindow display issues w/MultiUseWeapons.
        * importing melee and ranged weapons.
        * Gets acc and bulk working.  Passes tests.
        * Two weapon fighting on the move works.
        * Fixes tests.
        * Gets count and comparison work better with import.